Porting From Glewy
All of the files imported so far, with a few exceptions, have been taken from the Glewy git repo, in order to seperate these general purpose utility classes and functions that can be used in multiple projects, from the more specific graphics and game dev code within Glewy. Hopefully this should simplify the build tree of Glewy and allow substd to be developed seperately on its own.